+++ my-app/src/api/cardApi.ts
@@ -1,4 +1,3 @@
-
 export interface SendingData {
   number: string;
   expirationDate: string;
@@ -6,10 +5,19 @@ export interface SendingData {
   issuerCode: string;
 }
 
+export interface Card {
+  id: string;
+  issuerCode: string;
+  number: string;
+  expirationDate: string;
+}
+
+const API_BASE = import.meta.env.BASE_URL;
+
 export const postCard = async (
   sendingData: SendingData,
 ): Promise<{ id: string }> => {
-  const response = await fetch(`${import.meta.env.BASE_URL}cards`, {  
+  const response = await fetch(`${API_BASE}cards`, {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
     body: JSON.stringify(sendingData),
@@ -22,4 +30,22 @@ export const postCard = async (
   }
 
   return data;
+};
+
+export const getCards = async (): Promise<Card[]> => {
+  const response = await fetch(`${API_BASE}cards`);
+  if (!response.ok) {
+    throw new Error('카드 목록 불러오기 실패!');
+  }
+  return response.json();
+};
+
+export const deleteCard = async (id: string): Promise<void> => {
+  const response = await fetch(`${API_BASE}cards/${id}`, {
+    method: 'DELETE',
+  });
+
+  if (!response.ok) {
+    throw new Error('카드 삭제에 실패했습니다.');
+  }
 };
